Set the default stall mode to flush in all designs.
Catapult uses "stall" as the default stall mode, which means that if
input data is not available, the pipeline does not advance. This
simplifies the control logic but can "trap" valid data inside the
pipeline, leading to unexpected functional issues after HLS.

By setting the default stall mode to "flush", we are opting for reduced
odds of buggy RTL, at the possible cost of some area/power increase.

These pragmas can be overridden by HLS tcl directives if desired.